Bump deps

+++ wp-content/plugins/secure-custom-fields/includes/api/api-template.php
@@ -330,7 +330,7 @@ function acf_maybe_get_sub_field( $selectors, $post_id = false, $strict = true )
 	}
 
 	// vars
-	$offset    = acf_get_setting( 'row_index_offset' );
+	$offset    = (int) acf_get_setting( 'row_index_offset' );
 	$selector  = acf_extract_var( $selectors, 0 );
 	$selectors = array_values( $selectors ); // reset keys
 
@@ -346,7 +346,7 @@ function acf_maybe_get_sub_field( $selectors, $post_id = false, $strict = true )
 	for ( $j = 0; $j < count( $selectors ); $j += 2 ) {
 
 		// vars
-		$sub_i      = $selectors[ $j ];
+		$sub_i      = (int) $selectors[ $j ];
 		$sub_s      = $selectors[ $j + 1 ];
 		$field_name = $field['name'];
 
@@ -625,7 +625,7 @@ function have_rows( $selector, $post_id = false ) {
 	}
 
 	// Return true if next row exists.
-	if ( $active_loop && isset( $active_loop['value'][ $active_loop['i'] + 1 ] ) ) {
+	if ( $active_loop && isset( $active_loop['value'][ (int) $active_loop['i'] + 1 ] ) ) {
 		return true;
 	}
 
@@ -707,8 +707,8 @@ function get_row( $format = false ) {
 function get_row_index() {
 
 	// vars
-	$i      = acf_get_loop( 'active', 'i' );
-	$offset = acf_get_setting( 'row_index_offset' );
+	$i      = (int) acf_get_loop( 'active', 'i' );
+	$offset = (int) acf_get_setting( 'row_index_offset' );
 
 	// return
 	return $offset + $i;
@@ -1338,8 +1338,8 @@ function add_sub_row( $selector, $row = false, $post_id = false ) {
 function update_row( $selector, $i = 1, $row = false, $post_id = false ) {
 
 	// vars
-	$offset = acf_get_setting( 'row_index_offset' );
-	$i      = $i - $offset;
+	$offset = (int) acf_get_setting( 'row_index_offset' );
+	$i      = (int) $i - $offset;
 
 	// filter post_id
 	$post_id = acf_get_valid_post_id( $post_id );
@@ -1383,8 +1383,8 @@ function update_sub_row( $selector, $i = 1, $row = false, $post_id = false ) {
 
 	// vars
 	$sub_field = false;
-	$offset    = acf_get_setting( 'row_index_offset' );
-	$i         = $i - $offset;
+	$offset    = (int) acf_get_setting( 'row_index_offset' );
+	$i         = (int) $i - $offset;
 
 	// get sub field
 	if ( is_array( $selector ) ) {
@@ -1430,8 +1430,8 @@ function update_sub_row( $selector, $i = 1, $row = false, $post_id = false ) {
 function delete_row( $selector, $i = 1, $post_id = false ) {
 
 	// vars
-	$offset = acf_get_setting( 'row_index_offset' );
-	$i      = $i - $offset;
+	$offset = (int) acf_get_setting( 'row_index_offset' );
+	$i      = (int) $i - $offset;
 
 	// filter post_id
 	$post_id = acf_get_valid_post_id( $post_id );
@@ -1480,8 +1480,8 @@ function delete_sub_row( $selector, $i = 1, $post_id = false ) {
 
 	// vars
 	$sub_field = false;
-	$offset    = acf_get_setting( 'row_index_offset' );
-	$i         = $i - $offset;
+	$offset    = (int) acf_get_setting( 'row_index_offset' );
+	$i         = (int) $i - $offset;
 
 	// get sub field
 	if ( is_array( $selector ) ) {
